refactor(docs): move the interactable provider into runtimes/
the sample's provider is a docs runtime surface like docs, artifacts and playground, so it lives beside them and rides the two contract tests that already parameterize over every surface, instead of carrying a third copy of their mock harness.

drop the speech, dictation and feedback adapters: the sample renders no action bar and no mic button, so nothing can reach them. attachments stays, since ComposerPrimitive.Input attaches pasted images by default.
